Add a NewEngine function to create a new instance of the Engine struct
